+++ lib/nxt_support/serializers.rb
@@ -1,0 +1,1 @@
+require "nxt_support/serializers/has_time_attributes"